[mhlo] Fix warnings emited by -Wsign-compare
Building with clang (v14) produces numerous warnings about comparison
between signed and unsigned values. This patch introduces explicit cast
operations to eliminate the warnings. The unit tests run as part of the
`check-mlir-hlo` target pass.

+++ tensorflow/compiler/xla/mlir_hlo/lib/Dialect/gml_st/transforms/fusion_interface_impl.cc
@@ -82,7 +82,7 @@ struct LinalgGenericFusionInterface
         unsigned int rank = map.getNumResults();
         SmallVector<int64_t> permutation;
         permutation.reserve(rank);
-        for (int i = 0; i < rank; ++i) {
+        for (int i = 0; i < static_cast<int>(rank); ++i) {
           permutation.push_back(map.getPermutedPosition(i));
         }
         operandSubset = builder.create<TransposeDimsOp>(
